Move cluster creation prototype (#4034)
* Add link to Experiments page to cluster app creation form page

* Refactor some components

* Persist provider and branch selection in path

* Make some route params optional

* Use kebab case for provider name
