+++ src/components/MAPI/Experiments.tsx
@@ -1,7 +1,10 @@
 import { useAuthProvider } from 'Auth/MAPI/MapiAuthProvider';
 import DocumentTitle from 'components/shared/DocumentTitle';
 import { Box, Heading, Text } from 'grommet';
-import { AccountSettingsRoutes } from 'model/constants/routes';
+import {
+  AccountSettingsRoutes,
+  OrganizationsRoutes,
+} from 'model/constants/routes';
 import * as featureFlags from 'model/featureFlags';
 import { IAsynchronousDispatch } from 'model/stores/asynchronousAction';
 import {
@@ -9,10 +12,12 @@ import {
   setImpersonation,
 } from 'model/stores/main/actions';
 import { organizationsLoadMAPI } from 'model/stores/organization/actions';
+import { selectOrganizations } from 'model/stores/organization/selectors';
 import { IState } from 'model/stores/state';
 import React, { useState } from 'react';
 import { Breadcrumb } from 'react-breadcrumbs';
 import { useDispatch, useSelector } from 'react-redux';
+import { Link } from 'react-router-dom';
 import { useSWRConfig } from 'swr';
 import Button from 'UI/Controls/Button';
 import {
@@ -25,6 +30,7 @@ import {
 import CheckBoxInput from 'UI/Inputs/CheckBoxInput';
 import TextInput from 'UI/Inputs/TextInput';
 import { FlashMessage, messageTTL, messageType } from 'utils/flashMessage';
+import RoutePath from 'utils/routePath';
 
 import { usePermissionsKey } from './permissions/usePermissions';
 import { extractErrorMessage } from './utils';
@@ -124,6 +130,23 @@ const Experiments: React.FC<
     reloadOrganizations();
   };
 
+  const selectedOrgName = useSelector(
+    (state: IState) => state.main.selectedOrganization
+  );
+  const organizations = useSelector(selectOrganizations());
+  const selectedOrg = selectedOrgName
+    ? organizations[selectedOrgName]
+    : undefined;
+
+  const newClusterAppPath = selectedOrg
+    ? RoutePath.createUsablePath(
+        OrganizationsRoutes.Clusters.NewViaAppBundles,
+        {
+          orgId: selectedOrg.id,
+        }
+      )
+    : '';
+
   return (
     <Breadcrumb
       data={{
@@ -173,6 +196,28 @@ const Experiments: React.FC<
                   </TableCell>
                 </TableRow>
               ))}
+              <TableRow key='CreateClusterApp'>
+                <TableCell>
+                  <Link to={newClusterAppPath}>
+                    Create clusters via app bundles{' '}
+                    <i
+                      className='fa fa-open-in-new'
+                      role='presentation'
+                      aria-hidden={true}
+                    />
+                  </Link>
+                </TableCell>
+                <TableCell align='center' justify='center'>
+                  <CheckBoxInput
+                    toggle={true}
+                    margin='none'
+                    defaultChecked={true}
+                    name='CreateClusterApp'
+                    onChange={handleChange}
+                    disabled={true}
+                  />
+                </TableCell>
+              </TableRow>
             </TableBody>
           </Table>
         </Box>
