+++ src/model/constants/routes.ts
@@ -36,6 +36,9 @@ const OrganizationsRoutes = {
   Clusters: {
     Home: '/organizations/:orgId/clusters',
     New: '/organizations/:orgId/clusters/new',
+    NewViaAppBundles: '/organizations/:orgId/clusters/new-via-app-bundles',
+    NewViaAppBundlesProviderBranch:
+      '/organizations/:orgId/clusters/new-via-app-bundles/:provider/:branch',
     Detail: {
       Home: '/organizations/:orgId/clusters/:clusterId',
       KeyPairs: '/organizations/:orgId/clusters/:clusterId/keypairs',
